q3.scala

+++ q3.scala
@@ -3,10 +3,10 @@ object q3{
         num1 + num2
        }
      
-    def average(num1: Int,num2: Int): Double={
+    def average(num1: Int,num2: Int): BigDecimal={
         val num3 = add(num1,num2)
         val num4 = (num3/2)
-        BigDecimal(num4).setScale(2).toDouble                 // num4 rounded to two decimal places
+        BigDecimal(num4).setScale(2)                // num4 rounded to two decimal places
     }
 
     def main(args: Array[String]): Unit={
